requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-tinymce==4.0.0
idna==3.6
pillow==10.2.0
polib==1.2.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
